+++ demo/src/main/java/com/example/demo/repository/PersonRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface PersonRepository extends JpaRepository<Person,Integer>, CrudRepository<Person,Integer> {
+public interface PersonRepository extends JpaRepository<Person,Long>, CrudRepository<Person,Long> {
 
     Optional<Person> findById(Long id);
